@@ -1,415 +1,416 @@
-namespace ODTE.Optimization
+using ODTE.Strategy;
+
+namespace ODTE.Optimization;
+
+/// <summary>
+/// Genetic optimizer specifically for 24-day regime switching strategies using real historical data
+/// Optimizes parameters for regime classification, position sizing, and strategy selection
+/// </summary>
+public class RealDataRegimeOptimizer
 {
+    private readonly Random _random;
+    private readonly RealDataRegimeSwitcher _baseRegimeSwitcher;
+
+    public RealDataRegimeOptimizer()
+    {
+        _random = new Random(42); // Deterministic for reproducible results
+        _baseRegimeSwitcher = new RealDataRegimeSwitcher();
+    }
+
     /// <summary>
-    /// Genetic optimizer specifically for 24-day regime switching strategies using real historical data
-    /// Optimizes parameters for regime classification, position sizing, and strategy selection
+    /// Chromosome representing optimizable parameters for regime switching
     /// </summary>
-    public class RealDataRegimeOptimizer
+    public class RegimeChromosome
     {
-        private readonly Random _random;
-        private readonly RealDataRegimeSwitcher _baseRegimeSwitcher;
+        public double VixCalmThreshold { get; set; } = 20.0;      // VIX threshold for Calm regime
+        public double VixConvexThreshold { get; set; } = 35.0;    // VIX threshold for Convex regime
+        public double TrendCalmThreshold { get; set; } = 0.4;     // Trend score threshold for Calm
+        public double TrendConvexThreshold { get; set; } = 0.8;   // Trend score threshold for Convex
+        public double PositionSizingBase { get; set; } = 1.0;     // Base position sizing multiplier
+        public double CalmRegimeMultiplier { get; set; } = 1.2;   // Calm regime position multiplier
+        public double MixedRegimeMultiplier { get; set; } = 1.0;  // Mixed regime position multiplier
+        public double ConvexRegimeMultiplier { get; set; } = 0.7; // Convex regime position multiplier
+        public double StopLossMultiplier { get; set; } = 2.0;     // Stop loss threshold multiplier
+        public double ProfitTargetMultiplier { get; set; } = 0.5; // Profit target multiplier
 
-        public RealDataRegimeOptimizer()
+        // Fitness score
+        public double Fitness { get; set; }
+        public RegimeSwitchingAnalysisResult BacktestResult { get; set; }
+    }
+
+    /// <summary>
+    /// Run genetic optimization for specified number of generations
+    /// </summary>
+    public async Task<OptimizationResults> OptimizeAsync(int generations = 100, int populationSize = 50)
+    {
+        Console.WriteLine("🧬 GENETIC OPTIMIZATION: 24-Day Regime Switching on Real Data");
+        Console.WriteLine("==============================================================");
+        Console.WriteLine($"Target: {generations} generations with population of {populationSize}");
+        Console.WriteLine($"Data: Real SPY/VIX historical data (2015-2020)");
+        Console.WriteLine();
+
+        var results = new OptimizationResults
         {
-            _random = new Random(42); // Deterministic for reproducible results
-            _baseRegimeSwitcher = new RealDataRegimeSwitcher();
-        }
+            StartTime = DateTime.Now,
+            TargetGenerations = generations,
+            PopulationSize = populationSize,
+            GenerationHistory = new List<GenerationResult>()
+        };
 
-        /// <summary>
-        /// Chromosome representing optimizable parameters for regime switching
-        /// </summary>
-        public class RegimeChromosome
+        // Initialize population
+        var population = InitializePopulation(populationSize);
+
+        // Evolution loop
+        for (int generation = 0; generation < generations; generation++)
         {
-            public double VixCalmThreshold { get; set; } = 20.0;      // VIX threshold for Calm regime
-            public double VixConvexThreshold { get; set; } = 35.0;    // VIX threshold for Convex regime
-            public double TrendCalmThreshold { get; set; } = 0.4;     // Trend score threshold for Calm
-            public double TrendConvexThreshold { get; set; } = 0.8;   // Trend score threshold for Convex
-            public double PositionSizingBase { get; set; } = 1.0;     // Base position sizing multiplier
-            public double CalmRegimeMultiplier { get; set; } = 1.2;   // Calm regime position multiplier
-            public double MixedRegimeMultiplier { get; set; } = 1.0;  // Mixed regime position multiplier
-            public double ConvexRegimeMultiplier { get; set; } = 0.7; // Convex regime position multiplier
-            public double StopLossMultiplier { get; set; } = 2.0;     // Stop loss threshold multiplier
-            public double ProfitTargetMultiplier { get; set; } = 0.5; // Profit target multiplier
+            Console.WriteLine($"Generation {generation + 1}/{generations}");
 
-            // Fitness score
-            public double Fitness { get; set; }
-            public RegimeSwitchingAnalysisResult BacktestResult { get; set; }
-        }
+            // Evaluate fitness for all chromosomes
+            await EvaluatePopulationAsync(population);
 
-        /// <summary>
-        /// Run genetic optimization for specified number of generations
-        /// </summary>
-        public async Task<OptimizationResults> OptimizeAsync(int generations = 100, int populationSize = 50)
-        {
-            Console.WriteLine("🧬 GENETIC OPTIMIZATION: 24-Day Regime Switching on Real Data");
-            Console.WriteLine("==============================================================");
-            Console.WriteLine($"Target: {generations} generations with population of {populationSize}");
-            Console.WriteLine($"Data: Real SPY/VIX historical data (2015-2020)");
-            Console.WriteLine();
+            // Record generation statistics
+            var generationResult = AnalyzeGeneration(generation + 1, population);
+            results.GenerationHistory.Add(generationResult);
 
-            var results = new OptimizationResults
+            Console.WriteLine($"  Best Fitness: {generationResult.BestFitness:F2}");
+            Console.WriteLine($"  Avg Fitness: {generationResult.AvgFitness:F2}");
+            Console.WriteLine($"  Win Rate: {generationResult.BestWinRate:P1}");
+            Console.WriteLine($"  Total Return: {generationResult.BestTotalReturn:F1}%");
+
+            // Check for early termination on exceptional performance
+            if (generationResult.BestFitness > 1000.0)
             {
-                StartTime = DateTime.Now,
-                TargetGenerations = generations,
-                PopulationSize = populationSize,
-                GenerationHistory = new List<GenerationResult>()
+                Console.WriteLine($"🎯 Early termination: Exceptional fitness achieved at generation {generation + 1}");
+                break;
+            }
+
+            // Create next generation (skip on last iteration)
+            if (generation < generations - 1)
+            {
+                population = CreateNextGeneration(population);
+            }
+        }
+
+        results.EndTime = DateTime.Now;
+        results.BestChromosome = population.OrderByDescending(c => c.Fitness).First();
+
+        DisplayFinalResults(results);
+
+        return results;
+    }
+
+    private List<RegimeChromosome> InitializePopulation(int size)
+    {
+        var population = new List<RegimeChromosome>();
+
+        for (int i = 0; i < size; i++)
+        {
+            var chromosome = new RegimeChromosome
+            {
+                VixCalmThreshold = 15.0 + _random.NextDouble() * 15.0,      // 15-30
+                VixConvexThreshold = 25.0 + _random.NextDouble() * 25.0,    // 25-50
+                TrendCalmThreshold = 0.2 + _random.NextDouble() * 0.4,      // 0.2-0.6
+                TrendConvexThreshold = 0.6 + _random.NextDouble() * 0.4,    // 0.6-1.0
+                PositionSizingBase = 0.5 + _random.NextDouble() * 1.0,      // 0.5-1.5
+                CalmRegimeMultiplier = 0.8 + _random.NextDouble() * 0.8,    // 0.8-1.6
+                MixedRegimeMultiplier = 0.6 + _random.NextDouble() * 0.8,   // 0.6-1.4
+                ConvexRegimeMultiplier = 0.4 + _random.NextDouble() * 0.6,  // 0.4-1.0
+                StopLossMultiplier = 1.5 + _random.NextDouble() * 1.0,      // 1.5-2.5
+                ProfitTargetMultiplier = 0.3 + _random.NextDouble() * 0.4   // 0.3-0.7
             };
 
-            // Initialize population
-            var population = InitializePopulation(populationSize);
-
-            // Evolution loop
-            for (int generation = 0; generation < generations; generation++)
-            {
-                Console.WriteLine($"Generation {generation + 1}/{generations}");
-
-                // Evaluate fitness for all chromosomes
-                await EvaluatePopulationAsync(population);
-
-                // Record generation statistics
-                var generationResult = AnalyzeGeneration(generation + 1, population);
-                results.GenerationHistory.Add(generationResult);
-
-                Console.WriteLine($"  Best Fitness: {generationResult.BestFitness:F2}");
-                Console.WriteLine($"  Avg Fitness: {generationResult.AvgFitness:F2}");
-                Console.WriteLine($"  Win Rate: {generationResult.BestWinRate:P1}");
-                Console.WriteLine($"  Total Return: {generationResult.BestTotalReturn:F1}%");
-
-                // Check for early termination on exceptional performance
-                if (generationResult.BestFitness > 1000.0)
-                {
-                    Console.WriteLine($"🎯 Early termination: Exceptional fitness achieved at generation {generation + 1}");
-                    break;
-                }
-
-                // Create next generation (skip on last iteration)
-                if (generation < generations - 1)
-                {
-                    population = CreateNextGeneration(population);
-                }
-            }
-
-            results.EndTime = DateTime.Now;
-            results.BestChromosome = population.OrderByDescending(c => c.Fitness).First();
-
-            DisplayFinalResults(results);
-
-            return results;
+            population.Add(chromosome);
         }
 
-        private List<RegimeChromosome> InitializePopulation(int size)
+        return population;
+    }
+
+    private async Task EvaluatePopulationAsync(List<RegimeChromosome> population)
+    {
+        // Evaluate each chromosome in parallel for performance
+        var tasks = population.Select(EvaluateChromosomeAsync);
+        await Task.WhenAll(tasks);
+    }
+
+    private async Task EvaluateChromosomeAsync(RegimeChromosome chromosome)
+    {
+        try
         {
-            var population = new List<RegimeChromosome>();
+            // For now, use the base regime switcher and simulate parameter influence
+            // In a full implementation, we would need access to internal regime classification
+            var baseSwitcher = new RealDataRegimeSwitcher();
 
-            for (int i = 0; i < size; i++)
-            {
-                var chromosome = new RegimeChromosome
-                {
-                    VixCalmThreshold = 15.0 + _random.NextDouble() * 15.0,      // 15-30
-                    VixConvexThreshold = 25.0 + _random.NextDouble() * 25.0,    // 25-50
-                    TrendCalmThreshold = 0.2 + _random.NextDouble() * 0.4,      // 0.2-0.6
-                    TrendConvexThreshold = 0.6 + _random.NextDouble() * 0.4,    // 0.6-1.0
-                    PositionSizingBase = 0.5 + _random.NextDouble() * 1.0,      // 0.5-1.5
-                    CalmRegimeMultiplier = 0.8 + _random.NextDouble() * 0.8,    // 0.8-1.6
-                    MixedRegimeMultiplier = 0.6 + _random.NextDouble() * 0.8,   // 0.6-1.4
-                    ConvexRegimeMultiplier = 0.4 + _random.NextDouble() * 0.6,  // 0.4-1.0
-                    StopLossMultiplier = 1.5 + _random.NextDouble() * 1.0,      // 1.5-2.5
-                    ProfitTargetMultiplier = 0.3 + _random.NextDouble() * 0.4   // 0.3-0.7
-                };
+            // Run backtest on real historical data (2015-2020)
+            var result = baseSwitcher.RunRealDataAnalysis(
+                new DateTime(2015, 1, 1),
+                new DateTime(2020, 12, 31)
+            );
 
-                population.Add(chromosome);
-            }
+            // Apply chromosome-based modifications to the result
+            result = ModifyResultWithChromosome(result, chromosome);
 
-            return population;
+            chromosome.BacktestResult = result;
+
+            // Calculate fitness score (multi-objective optimization)
+            chromosome.Fitness = CalculateFitness(result);
         }
-
-        private async Task EvaluatePopulationAsync(List<RegimeChromosome> population)
+        catch (Exception ex)
         {
-            // Evaluate each chromosome in parallel for performance
-            var tasks = population.Select(EvaluateChromosomeAsync);
-            await Task.WhenAll(tasks);
-        }
-
-        private async Task EvaluateChromosomeAsync(RegimeChromosome chromosome)
-        {
-            try
-            {
-                // For now, use the base regime switcher and simulate parameter influence
-                // In a full implementation, we would need access to internal regime classification
-                var baseSwitcher = new RealDataRegimeSwitcher();
-
-                // Run backtest on real historical data (2015-2020)
-                var result = baseSwitcher.RunRealDataAnalysis(
-                    new DateTime(2015, 1, 1),
-                    new DateTime(2020, 12, 31)
-                );
-
-                // Apply chromosome-based modifications to the result
-                result = ModifyResultWithChromosome(result, chromosome);
-
-                chromosome.BacktestResult = result;
-
-                // Calculate fitness score (multi-objective optimization)
-                chromosome.Fitness = CalculateFitness(result);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Warning: Chromosome evaluation failed: {ex.Message}");
-                chromosome.Fitness = -1000.0; // Penalty for failed evaluation
-            }
-        }
-
-        private RegimeSwitchingAnalysisResult ModifyResultWithChromosome(
-            RegimeSwitchingAnalysisResult baseResult,
-            RegimeChromosome chromosome)
-        {
-            // Simulate the effect of different parameters on the strategy performance
-            var modifiedResult = new RegimeSwitchingAnalysisResult
-            {
-                Periods = baseResult.Periods,
-                TotalPeriods = baseResult.TotalPeriods,
-                RegimePerformance = baseResult.RegimePerformance
-            };
-
-            // Adjust performance based on chromosome parameters
-            double performanceMultiplier = CalculatePerformanceMultiplier(chromosome);
-
-            modifiedResult.AverageReturn = baseResult.AverageReturn * performanceMultiplier;
-            modifiedResult.TotalReturn = Math.Pow(1 + baseResult.AverageReturn * performanceMultiplier / 100, baseResult.TotalPeriods) * 100 - 100;
-            modifiedResult.BestPeriodReturn = baseResult.BestPeriodReturn * Math.Max(0.8, performanceMultiplier);
-            modifiedResult.WorstPeriodReturn = baseResult.WorstPeriodReturn * Math.Min(1.2, performanceMultiplier);
-
-            // Adjust win rate based on parameter quality
-            var winRateAdjustment = CalculateWinRateAdjustment(chromosome);
-            modifiedResult.WinRate = Math.Max(0.3, Math.Min(0.95, baseResult.WinRate + winRateAdjustment));
-
-            return modifiedResult;
-        }
-
-        private double CalculatePerformanceMultiplier(RegimeChromosome chromosome)
-        {
-            // Score based on parameter reasonableness
-            double score = 1.0;
-
-            // VIX thresholds should be reasonable
-            if (chromosome.VixCalmThreshold > 10 && chromosome.VixCalmThreshold < 25)
-                score += 0.1;
-            if (chromosome.VixConvexThreshold > 30 && chromosome.VixConvexThreshold < 50)
-                score += 0.1;
-            if (chromosome.VixConvexThreshold > chromosome.VixCalmThreshold + 5)
-                score += 0.1;
-
-            // Position sizing should be conservative
-            if (chromosome.PositionSizingBase > 0.7 && chromosome.PositionSizingBase < 1.3)
-                score += 0.1;
-            if (chromosome.CalmRegimeMultiplier > chromosome.ConvexRegimeMultiplier)
-                score += 0.1; // More aggressive in calm markets
-
-            // Add some randomness for genetic diversity
-            score += (_random.NextDouble() - 0.5) * 0.2;
-
-            return Math.Max(0.5, Math.Min(1.5, score));
-        }
-
-        private double CalculateWinRateAdjustment(RegimeChromosome chromosome)
-        {
-            // Adjust win rate based on parameter quality
-            double adjustment = 0.0;
-
-            // Better thresholds should improve win rate
-            if (chromosome.VixCalmThreshold > 15 && chromosome.VixCalmThreshold < 22)
-                adjustment += 0.02;
-            if (chromosome.TrendCalmThreshold > 0.3 && chromosome.TrendCalmThreshold < 0.5)
-                adjustment += 0.02;
-            if (chromosome.StopLossMultiplier > 1.8 && chromosome.StopLossMultiplier < 2.2)
-                adjustment += 0.01;
-
-            // Add randomness
-            adjustment += (_random.NextDouble() - 0.5) * 0.05;
-
-            return adjustment;
-        }
-
-        private double CalculateFitness(RegimeSwitchingAnalysisResult result)
-        {
-            if (result == null || result.TotalPeriods == 0)
-                return -1000.0;
-
-            // Multi-objective fitness function
-            var winRate = result.WinRate;
-            var avgReturn = result.AverageReturn;
-            var totalReturn = result.TotalReturn;
-            var sharpeApprox = avgReturn / Math.Max(1.0, Math.Abs(result.WorstPeriodReturn)); // Simplified Sharpe
-
-            // Weighted fitness score
-            var fitness =
-                (winRate * 200.0) +                    // Win rate (max 200 points)
-                (avgReturn * 10.0) +                   // Average return (up to ~100 points)
-                (Math.Log10(Math.Max(1, totalReturn)) * 50.0) + // Total return (logarithmic scaling)
-                (sharpeApprox * 30.0) +                // Risk-adjusted return
-                (result.TotalPeriods * 2.0);           // Bonus for completing more periods
-
-            // Penalty for poor performance
-            if (winRate < 0.5) fitness -= 100.0;
-            if (avgReturn < 0) fitness -= 200.0;
-            if (result.WorstPeriodReturn < -20) fitness -= 150.0; // Large drawdown penalty
-
-            return fitness;
-        }
-
-        private List<RegimeChromosome> CreateNextGeneration(List<RegimeChromosome> currentGen)
-        {
-            var sortedPop = currentGen.OrderByDescending(c => c.Fitness).ToList();
-            var nextGen = new List<RegimeChromosome>();
-
-            // Elite selection (keep top 20%)
-            int eliteCount = Math.Max(1, currentGen.Count / 5);
-            nextGen.AddRange(sortedPop.Take(eliteCount));
-
-            // Generate offspring to fill population
-            while (nextGen.Count < currentGen.Count)
-            {
-                var parent1 = TournamentSelection(sortedPop);
-                var parent2 = TournamentSelection(sortedPop);
-
-                var offspring = Crossover(parent1, parent2);
-                Mutate(offspring);
-
-                nextGen.Add(offspring);
-            }
-
-            return nextGen;
-        }
-
-        private RegimeChromosome TournamentSelection(List<RegimeChromosome> population)
-        {
-            const int tournamentSize = 5;
-            var tournament = new List<RegimeChromosome>();
-
-            for (int i = 0; i < tournamentSize; i++)
-            {
-                var randomIndex = _random.Next(population.Count);
-                tournament.Add(population[randomIndex]);
-            }
-
-            return tournament.OrderByDescending(c => c.Fitness).First();
-        }
-
-        private RegimeChromosome Crossover(RegimeChromosome parent1, RegimeChromosome parent2)
-        {
-            return new RegimeChromosome
-            {
-                VixCalmThreshold = _random.NextDouble() < 0.5 ? parent1.VixCalmThreshold : parent2.VixCalmThreshold,
-                VixConvexThreshold = _random.NextDouble() < 0.5 ? parent1.VixConvexThreshold : parent2.VixConvexThreshold,
-                TrendCalmThreshold = _random.NextDouble() < 0.5 ? parent1.TrendCalmThreshold : parent2.TrendCalmThreshold,
-                TrendConvexThreshold = _random.NextDouble() < 0.5 ? parent1.TrendConvexThreshold : parent2.TrendConvexThreshold,
-                PositionSizingBase = _random.NextDouble() < 0.5 ? parent1.PositionSizingBase : parent2.PositionSizingBase,
-                CalmRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.CalmRegimeMultiplier : parent2.CalmRegimeMultiplier,
-                MixedRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.MixedRegimeMultiplier : parent2.MixedRegimeMultiplier,
-                ConvexRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.ConvexRegimeMultiplier : parent2.ConvexRegimeMultiplier,
-                StopLossMultiplier = _random.NextDouble() < 0.5 ? parent1.StopLossMultiplier : parent2.StopLossMultiplier,
-                ProfitTargetMultiplier = _random.NextDouble() < 0.5 ? parent1.ProfitTargetMultiplier : parent2.ProfitTargetMultiplier
-            };
-        }
-
-        private void Mutate(RegimeChromosome chromosome)
-        {
-            const double mutationRate = 0.1;
-            const double mutationStrength = 0.1;
-
-            if (_random.NextDouble() < mutationRate)
-                chromosome.VixCalmThreshold *= (1.0 + (_random.NextDouble() - 0.5) * mutationStrength);
-
-            if (_random.NextDouble() < mutationRate)
-                chromosome.VixConvexThreshold *= (1.0 + (_random.NextDouble() - 0.5) * mutationStrength);
-
-            if (_random.NextDouble() < mutationRate)
-                chromosome.TrendCalmThreshold = Math.Max(0.1, Math.Min(0.9,
-                    chromosome.TrendCalmThreshold + (_random.NextDouble() - 0.5) * 0.1));
-
-            if (_random.NextDouble() < mutationRate)
-                chromosome.TrendConvexThreshold = Math.Max(0.5, Math.Min(1.0,
-                    chromosome.TrendConvexThreshold + (_random.NextDouble() - 0.5) * 0.1));
-
-            if (_random.NextDouble() < mutationRate)
-                chromosome.PositionSizingBase = Math.Max(0.2, Math.Min(2.0,
-                    chromosome.PositionSizingBase * (1.0 + (_random.NextDouble() - 0.5) * mutationStrength)));
-
-            // Additional mutations for other parameters...
-        }
-
-        private GenerationResult AnalyzeGeneration(int generation, List<RegimeChromosome> population)
-        {
-            var validPop = population.Where(c => c.Fitness > -1000).ToList();
-            if (!validPop.Any()) validPop = population; // Fallback if all failed
-
-            var best = validPop.OrderByDescending(c => c.Fitness).First();
-
-            return new GenerationResult
-            {
-                GenerationNumber = generation,
-                BestFitness = best.Fitness,
-                AvgFitness = validPop.Average(c => c.Fitness),
-                BestWinRate = best.BacktestResult?.WinRate ?? 0,
-                BestTotalReturn = best.BacktestResult?.TotalReturn ?? 0,
-                BestChromosome = best
-            };
-        }
-
-        private void DisplayFinalResults(OptimizationResults results)
-        {
-            Console.WriteLine();
-            Console.WriteLine("🏆 GENETIC OPTIMIZATION COMPLETED");
-            Console.WriteLine("===================================");
-            Console.WriteLine($"Duration: {(results.EndTime - results.StartTime).TotalMinutes:F1} minutes");
-            Console.WriteLine($"Generations: {results.GenerationHistory.Count}");
-            Console.WriteLine();
-
-            var best = results.BestChromosome;
-            Console.WriteLine("🥇 BEST CHROMOSOME PARAMETERS:");
-            Console.WriteLine($"  VIX Calm Threshold: {best.VixCalmThreshold:F2}");
-            Console.WriteLine($"  VIX Convex Threshold: {best.VixConvexThreshold:F2}");
-            Console.WriteLine($"  Trend Calm Threshold: {best.TrendCalmThreshold:F3}");
-            Console.WriteLine($"  Trend Convex Threshold: {best.TrendConvexThreshold:F3}");
-            Console.WriteLine($"  Position Sizing Base: {best.PositionSizingBase:F2}");
-            Console.WriteLine($"  Calm Regime Multiplier: {best.CalmRegimeMultiplier:F2}");
-            Console.WriteLine($"  Mixed Regime Multiplier: {best.MixedRegimeMultiplier:F2}");
-            Console.WriteLine($"  Convex Regime Multiplier: {best.ConvexRegimeMultiplier:F2}");
-            Console.WriteLine();
-
-            if (best.BacktestResult != null)
-            {
-                Console.WriteLine("📊 OPTIMIZED PERFORMANCE:");
-                Console.WriteLine($"  Fitness Score: {best.Fitness:F2}");
-                Console.WriteLine($"  Win Rate: {best.BacktestResult.WinRate:P1}");
-                Console.WriteLine($"  Average Return: {best.BacktestResult.AverageReturn:F2}%");
-                Console.WriteLine($"  Total Return: {best.BacktestResult.TotalReturn:F1}%");
-                Console.WriteLine($"  Best Period: {best.BacktestResult.BestPeriodReturn:F2}%");
-                Console.WriteLine($"  Worst Period: {best.BacktestResult.WorstPeriodReturn:F2}%");
-                Console.WriteLine($"  Total Periods: {best.BacktestResult.TotalPeriods}");
-            }
+            Console.WriteLine($"Warning: Chromosome evaluation failed: {ex.Message}");
+            chromosome.Fitness = -1000.0; // Penalty for failed evaluation
         }
     }
 
-
-    // Results data structures
-    public class OptimizationResults
+    private RegimeSwitchingAnalysisResult ModifyResultWithChromosome(
+        RegimeSwitchingAnalysisResult baseResult,
+        RegimeChromosome chromosome)
     {
-        public DateTime StartTime { get; set; }
-        public DateTime EndTime { get; set; }
-        public int TargetGenerations { get; set; }
-        public int PopulationSize { get; set; }
-        public List<GenerationResult> GenerationHistory { get; set; }
-        public RealDataRegimeOptimizer.RegimeChromosome BestChromosome { get; set; }
+        // Simulate the effect of different parameters on the strategy performance
+        var modifiedResult = new RegimeSwitchingAnalysisResult
+        {
+            Periods = baseResult.Periods,
+            TotalPeriods = baseResult.TotalPeriods,
+            RegimePerformance = baseResult.RegimePerformance
+        };
+
+        // Adjust performance based on chromosome parameters
+        double performanceMultiplier = CalculatePerformanceMultiplier(chromosome);
+
+        modifiedResult.AverageReturn = baseResult.AverageReturn * performanceMultiplier;
+        modifiedResult.TotalReturn = Math.Pow(1 + baseResult.AverageReturn * performanceMultiplier / 100, baseResult.TotalPeriods) * 100 - 100;
+        modifiedResult.BestPeriodReturn = baseResult.BestPeriodReturn * Math.Max(0.8, performanceMultiplier);
+        modifiedResult.WorstPeriodReturn = baseResult.WorstPeriodReturn * Math.Min(1.2, performanceMultiplier);
+
+        // Adjust win rate based on parameter quality
+        var winRateAdjustment = CalculateWinRateAdjustment(chromosome);
+        modifiedResult.WinRate = Math.Max(0.3, Math.Min(0.95, baseResult.WinRate + winRateAdjustment));
+
+        return modifiedResult;
     }
 
-    public class GenerationResult
+    private double CalculatePerformanceMultiplier(RegimeChromosome chromosome)
     {
-        public int GenerationNumber { get; set; }
-        public double BestFitness { get; set; }
-        public double AvgFitness { get; set; }
-        public double BestWinRate { get; set; }
-        public double BestTotalReturn { get; set; }
-        public RealDataRegimeOptimizer.RegimeChromosome BestChromosome { get; set; }
+        // Score based on parameter reasonableness
+        double score = 1.0;
+
+        // VIX thresholds should be reasonable
+        if (chromosome.VixCalmThreshold > 10 && chromosome.VixCalmThreshold < 25)
+            score += 0.1;
+        if (chromosome.VixConvexThreshold > 30 && chromosome.VixConvexThreshold < 50)
+            score += 0.1;
+        if (chromosome.VixConvexThreshold > chromosome.VixCalmThreshold + 5)
+            score += 0.1;
+
+        // Position sizing should be conservative
+        if (chromosome.PositionSizingBase > 0.7 && chromosome.PositionSizingBase < 1.3)
+            score += 0.1;
+        if (chromosome.CalmRegimeMultiplier > chromosome.ConvexRegimeMultiplier)
+            score += 0.1; // More aggressive in calm markets
+
+        // Add some randomness for genetic diversity
+        score += (_random.NextDouble() - 0.5) * 0.2;
+
+        return Math.Max(0.5, Math.Min(1.5, score));
     }
+
+    private double CalculateWinRateAdjustment(RegimeChromosome chromosome)
+    {
+        // Adjust win rate based on parameter quality
+        double adjustment = 0.0;
+
+        // Better thresholds should improve win rate
+        if (chromosome.VixCalmThreshold > 15 && chromosome.VixCalmThreshold < 22)
+            adjustment += 0.02;
+        if (chromosome.TrendCalmThreshold > 0.3 && chromosome.TrendCalmThreshold < 0.5)
+            adjustment += 0.02;
+        if (chromosome.StopLossMultiplier > 1.8 && chromosome.StopLossMultiplier < 2.2)
+            adjustment += 0.01;
+
+        // Add randomness
+        adjustment += (_random.NextDouble() - 0.5) * 0.05;
+
+        return adjustment;
+    }
+
+    private double CalculateFitness(RegimeSwitchingAnalysisResult result)
+    {
+        if (result == null || result.TotalPeriods == 0)
+            return -1000.0;
+
+        // Multi-objective fitness function
+        var winRate = result.WinRate;
+        var avgReturn = result.AverageReturn;
+        var totalReturn = result.TotalReturn;
+        var sharpeApprox = avgReturn / Math.Max(1.0, Math.Abs(result.WorstPeriodReturn)); // Simplified Sharpe
+
+        // Weighted fitness score
+        var fitness =
+            (winRate * 200.0) +                    // Win rate (max 200 points)
+            (avgReturn * 10.0) +                   // Average return (up to ~100 points)
+            (Math.Log10(Math.Max(1, totalReturn)) * 50.0) + // Total return (logarithmic scaling)
+            (sharpeApprox * 30.0) +                // Risk-adjusted return
+            (result.TotalPeriods * 2.0);           // Bonus for completing more periods
+
+        // Penalty for poor performance
+        if (winRate < 0.5) fitness -= 100.0;
+        if (avgReturn < 0) fitness -= 200.0;
+        if (result.WorstPeriodReturn < -20) fitness -= 150.0; // Large drawdown penalty
+
+        return fitness;
+    }
+
+    private List<RegimeChromosome> CreateNextGeneration(List<RegimeChromosome> currentGen)
+    {
+        var sortedPop = currentGen.OrderByDescending(c => c.Fitness).ToList();
+        var nextGen = new List<RegimeChromosome>();
+
+        // Elite selection (keep top 20%)
+        int eliteCount = Math.Max(1, currentGen.Count / 5);
+        nextGen.AddRange(sortedPop.Take(eliteCount));
+
+        // Generate offspring to fill population
+        while (nextGen.Count < currentGen.Count)
+        {
+            var parent1 = TournamentSelection(sortedPop);
+            var parent2 = TournamentSelection(sortedPop);
+
+            var offspring = Crossover(parent1, parent2);
+            Mutate(offspring);
+
+            nextGen.Add(offspring);
+        }
+
+        return nextGen;
+    }
+
+    private RegimeChromosome TournamentSelection(List<RegimeChromosome> population)
+    {
+        const int tournamentSize = 5;
+        var tournament = new List<RegimeChromosome>();
+
+        for (int i = 0; i < tournamentSize; i++)
+        {
+            var randomIndex = _random.Next(population.Count);
+            tournament.Add(population[randomIndex]);
+        }
+
+        return tournament.OrderByDescending(c => c.Fitness).First();
+    }
+
+    private RegimeChromosome Crossover(RegimeChromosome parent1, RegimeChromosome parent2)
+    {
+        return new RegimeChromosome
+        {
+            VixCalmThreshold = _random.NextDouble() < 0.5 ? parent1.VixCalmThreshold : parent2.VixCalmThreshold,
+            VixConvexThreshold = _random.NextDouble() < 0.5 ? parent1.VixConvexThreshold : parent2.VixConvexThreshold,
+            TrendCalmThreshold = _random.NextDouble() < 0.5 ? parent1.TrendCalmThreshold : parent2.TrendCalmThreshold,
+            TrendConvexThreshold = _random.NextDouble() < 0.5 ? parent1.TrendConvexThreshold : parent2.TrendConvexThreshold,
+            PositionSizingBase = _random.NextDouble() < 0.5 ? parent1.PositionSizingBase : parent2.PositionSizingBase,
+            CalmRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.CalmRegimeMultiplier : parent2.CalmRegimeMultiplier,
+            MixedRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.MixedRegimeMultiplier : parent2.MixedRegimeMultiplier,
+            ConvexRegimeMultiplier = _random.NextDouble() < 0.5 ? parent1.ConvexRegimeMultiplier : parent2.ConvexRegimeMultiplier,
+            StopLossMultiplier = _random.NextDouble() < 0.5 ? parent1.StopLossMultiplier : parent2.StopLossMultiplier,
+            ProfitTargetMultiplier = _random.NextDouble() < 0.5 ? parent1.ProfitTargetMultiplier : parent2.ProfitTargetMultiplier
+        };
+    }
+
+    private void Mutate(RegimeChromosome chromosome)
+    {
+        const double mutationRate = 0.1;
+        const double mutationStrength = 0.1;
+
+        if (_random.NextDouble() < mutationRate)
+            chromosome.VixCalmThreshold *= (1.0 + (_random.NextDouble() - 0.5) * mutationStrength);
+
+        if (_random.NextDouble() < mutationRate)
+            chromosome.VixConvexThreshold *= (1.0 + (_random.NextDouble() - 0.5) * mutationStrength);
+
+        if (_random.NextDouble() < mutationRate)
+            chromosome.TrendCalmThreshold = Math.Max(0.1, Math.Min(0.9,
+                chromosome.TrendCalmThreshold + (_random.NextDouble() - 0.5) * 0.1));
+
+        if (_random.NextDouble() < mutationRate)
+            chromosome.TrendConvexThreshold = Math.Max(0.5, Math.Min(1.0,
+                chromosome.TrendConvexThreshold + (_random.NextDouble() - 0.5) * 0.1));
+
+        if (_random.NextDouble() < mutationRate)
+            chromosome.PositionSizingBase = Math.Max(0.2, Math.Min(2.0,
+                chromosome.PositionSizingBase * (1.0 + (_random.NextDouble() - 0.5) * mutationStrength)));
+
+        // Additional mutations for other parameters...
+    }
+
+    private GenerationResult AnalyzeGeneration(int generation, List<RegimeChromosome> population)
+    {
+        var validPop = population.Where(c => c.Fitness > -1000).ToList();
+        if (!validPop.Any()) validPop = population; // Fallback if all failed
+
+        var best = validPop.OrderByDescending(c => c.Fitness).First();
+
+        return new GenerationResult
+        {
+            GenerationNumber = generation,
+            BestFitness = best.Fitness,
+            AvgFitness = validPop.Average(c => c.Fitness),
+            BestWinRate = best.BacktestResult?.WinRate ?? 0,
+            BestTotalReturn = best.BacktestResult?.TotalReturn ?? 0,
+            BestChromosome = best
+        };
+    }
+
+    private void DisplayFinalResults(OptimizationResults results)
+    {
+        Console.WriteLine();
+        Console.WriteLine("🏆 GENETIC OPTIMIZATION COMPLETED");
+        Console.WriteLine("===================================");
+        Console.WriteLine($"Duration: {(results.EndTime - results.StartTime).TotalMinutes:F1} minutes");
+        Console.WriteLine($"Generations: {results.GenerationHistory.Count}");
+        Console.WriteLine();
+
+        var best = results.BestChromosome;
+        Console.WriteLine("🥇 BEST CHROMOSOME PARAMETERS:");
+        Console.WriteLine($"  VIX Calm Threshold: {best.VixCalmThreshold:F2}");
+        Console.WriteLine($"  VIX Convex Threshold: {best.VixConvexThreshold:F2}");
+        Console.WriteLine($"  Trend Calm Threshold: {best.TrendCalmThreshold:F3}");
+        Console.WriteLine($"  Trend Convex Threshold: {best.TrendConvexThreshold:F3}");
+        Console.WriteLine($"  Position Sizing Base: {best.PositionSizingBase:F2}");
+        Console.WriteLine($"  Calm Regime Multiplier: {best.CalmRegimeMultiplier:F2}");
+        Console.WriteLine($"  Mixed Regime Multiplier: {best.MixedRegimeMultiplier:F2}");
+        Console.WriteLine($"  Convex Regime Multiplier: {best.ConvexRegimeMultiplier:F2}");
+        Console.WriteLine();
+
+        if (best.BacktestResult != null)
+        {
+            Console.WriteLine("📊 OPTIMIZED PERFORMANCE:");
+            Console.WriteLine($"  Fitness Score: {best.Fitness:F2}");
+            Console.WriteLine($"  Win Rate: {best.BacktestResult.WinRate:P1}");
+            Console.WriteLine($"  Average Return: {best.BacktestResult.AverageReturn:F2}%");
+            Console.WriteLine($"  Total Return: {best.BacktestResult.TotalReturn:F1}%");
+            Console.WriteLine($"  Best Period: {best.BacktestResult.BestPeriodReturn:F2}%");
+            Console.WriteLine($"  Worst Period: {best.BacktestResult.WorstPeriodReturn:F2}%");
+            Console.WriteLine($"  Total Periods: {best.BacktestResult.TotalPeriods}");
+        }
+    }
+}
+
+
+// Results data structures
+public class OptimizationResults
+{
+    public DateTime StartTime { get; set; }
+    public DateTime EndTime { get; set; }
+    public int TargetGenerations { get; set; }
+    public int PopulationSize { get; set; }
+    public List<GenerationResult> GenerationHistory { get; set; }
+    public RealDataRegimeOptimizer.RegimeChromosome BestChromosome { get; set; }
+}
+
+public class GenerationResult
+{
+    public int GenerationNumber { get; set; }
+    public double BestFitness { get; set; }
+    public double AvgFitness { get; set; }
+    public double BestWinRate { get; set; }
+    public double BestTotalReturn { get; set; }
+    public RealDataRegimeOptimizer.RegimeChromosome BestChromosome { get; set; }
 }
